+++ src/main/java/com/malikov/shopsystem/repository/OrderRepository.java
@@ -8,9 +8,9 @@ import java.util.List;
 
 public interface OrderRepository extends Repository<Order> {
 
-    Collection<Order> getByCustomerId(int customerId);
+    Collection<Order> getByCustomerId(Long customerId);
 
-    Collection<Order> getByProductId(int productId);
+    Collection<Order> getByProductId(Long productId);
 
     void updateStatus(Integer orderId, OrderStatus status);
