+++ src/main/java/com/malikov/shopsystem/service/ProductService.java
@@ -6,13 +6,14 @@ import java.util.Collection;
 import java.util.List;
 
 public interface ProductService extends Service<Product> {
-    Collection<Product> getByCategoryId(int categoryId);
+
+    Collection<Product> getByCategoryId(Long categoryId);
 
     Collection<Product> getAllQuantityLessThan(int quantity);
 
-    void enableUnlimited(int id, boolean unlimited);
+    void enableUnlimited(Long id, boolean unlimited);
 
-    void enableHasVariations(int id, boolean hasVariations);
+    void enableHasVariations(Long id, boolean hasVariations);
 
     List<Product> getByProductNameMask(String productNameMask);
