+++ src/main/java/com/malikov/shopsystem/to/OrderItemAutocompleteTo.java
@@ -5,15 +5,15 @@ public class OrderItemAutocompleteTo {
    // TODO: 3/25/2017    get rid of label. use javascript concatenation instead
     private String label;
 
-    private Integer productId;
+    private Long productId;
 
-    private Integer productVariationId;
+    private Long productVariationId;
 
     private String orderItemName;
 
     private Integer orderItemPrice;
 
-    public OrderItemAutocompleteTo(String label, Integer productId, Integer productVariationId, String orderItemName, Integer orderItemPrice) {
+    public OrderItemAutocompleteTo(String label, Long productId, Long productVariationId, String orderItemName, Integer orderItemPrice) {
         this.label = label;
         this.productId = productId;
         this.productVariationId = productVariationId;
@@ -29,19 +29,19 @@ public class OrderItemAutocompleteTo {
         this.label = label;
     }
 
-    public Integer getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(Integer productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
 
-    public Integer getProductVariationId() {
+    public Long getProductVariationId() {
         return productVariationId;
     }
 
-    public void setProductVariationId(Integer productVariationId) {
+    public void setProductVariationId(Long productVariationId) {
         this.productVariationId = productVariationId;
     }
